back-end/requirements.txt
# Flask 및 확장 모듈
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Marshmallow==0.15.0
Flask-CORS==4.0.0

# 데이터베이스 드라이버
PyMySQL==1.1.0

# 유틸리티
python-dotenv==1.0.0
marshmallow-sqlalchemy==0.30.0
Werkzeug==2.3.7

# 이미지 처리 (선택적)
Pillow==10.1.0

# 서버 (프로덕션용, 선택적)
gunicorn==21.2.0